Move test models factory to test_app
