+++ mapentity/tests/test_attachments.py
@@ -1,5 +1,4 @@
 from unittest import mock
-from django.core.management import call_command
 
 from django.test import TestCase, RequestFactory
 from django.contrib.auth import get_user_model
@@ -11,7 +10,7 @@ from django.urls import reverse
 from paperclip.settings import get_attachment_model, get_filetype_model
 from mapentity.views.generic import MapEntityDetail
 
-from mapentity.tests.factories import DummyModelFactory
+from test_app.factories import DummyModelFactory
 from test_app.models import DummyModel
